requirements.txt
Django>=3.0,<4.0.2
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.1

# flake8>=3.6.0,<3.7.0